Update README.md
This report is unique to Hopper Division, in that each Category had Subcategories under it when entering a delay. Due to the nature of the Mate being responsible for the safety of the vessel, and frequently in shallow waters, a subcategory of “Other” was the default.  The intention was for either the Mate, Captain, Chief, or Site Engineer would correct the data to more appropriate at a time deemed safe to try to capture that asset level of fault; however, this was not the case most of the time.
I created this report to help on a weekly basis to determine incorrect Other subcategories and to address them, and as corrected would eventually fall off the report. Again, the raw EDL data is the source, since it is the most comprehensive way of now looking at delays during times of production. Semi-automating the database down and finally data-mining this into an easier and more rapid way of addressing these at a weekly management meeting for that division.